Big refactor of how patches are organized and loaded. Conditionally loads optional patches now. Removed hard dependencies to MoreCompany and 3rdPerson.
