+++ .gitignore
@@ -7,8 +7,8 @@
 Libraries/*
 *.dll
 *.zip
-Distribution/icon.xcf
-Distribution/README.md
+Dist/icon.xcf
+Dist/README.md
 
 # User-specific files
 *.rsuser
